Refactor project structure for better organization + Add functions for recognize_clothe_and_color.py + make one endpoint
