+++ config.py
@@ -1,0 +1,7 @@
+import cloudinary
+          
+cloudinary.config( 
+  cloud_name = "depgto6ws", 
+  api_key = "167399317712469", 
+  api_secret = "WPdHcGvOFk2jfpu61ahzLaGkCW4" 
+)
