+++ output_image.jpg
@@ -1,0 +1,1 @@
+{"image_without_background_url":"https://res.cloudinary.com/depgto6ws/image/upload/v1716570666/6b84980a-9e50-443e-a4e7-a86a2d77eed3.png","label":"Watche","color_palette":["#b4b1aa","#1e2a14","#58422b","#79643d","#636c67"]}
